fix: separate out building where clause from creating filter schema to prevent importing sequelize into client
